A readable stream should emit end when it ends
The main purpose for this is so that pipe gets the end event of the readable stream
and can clean up the event listeners otherwise memory leak hell

The problem is that it breaks test/index.js and I dont know why.
Better yet I don't understand your test. You will have to fix this. Sorry!